requirements.txt
cryptography==42.0.5
google_api_python_client==2.127.0
google_auth_oauthlib==1.2.0
Requests==2.31.0
